requirements.txt
requests==2.28.1
fastapi==0.95.1
uvicorn==0.22.0
PyYAML==6.0
